BOM RD완료, CU (ing)

+++ mealkit-project/src/main/java/org/mealkitspringboot/domain/BomListVo.java
@@ -6,20 +6,20 @@ import lombok.Setter;
 @Getter
 @Setter
 public class BomListVo {
-	private int listSeq;
-	private int bomId;
-	private int prodId;
+	private Long listSeq;
+	private Long bomId;
+	private Long prodId;
 	private String prodNm;
 	private String prodDiv;
 	private String prodSpec;
-	private int prodPrice;
-	private int lotId;
-	private int lotSize;
+	private Long prodPrice;
+	private Long lotId;
+	private Long lotSize;
 	private String matDiv;
-	private int matId;
+	private Long matId;
 	private String matNm;
 	private String quantityUnits;
-	private int bomProdQuantity;
+	private Long bomProdQuantity;
